+++ rhine-gloss/src/FRP/Rhine/Gloss/Internals.hs
@@ -36,9 +36,9 @@ errMsg =  "You cannot start gloss apps with FRP.Rhine.flow. "
 data GlossEventClock = GlossEventClock
 
 instance Clock m GlossEventClock where
-  type TimeDomainOf GlossEventClock = ()
-  type Tag          GlossEventClock = Event
-  startClock _ = error errMsg
+  type Time GlossEventClock = ()
+  type Tag  GlossEventClock = Event
+  initClock _ = error errMsg
 
 -- | The clock that ticks for every @gloss@ simulation step,
 --   but only shows the time delta in the tag.
@@ -46,43 +46,43 @@ instance Clock m GlossEventClock where
 data GlossSimulationClock_ = GlossSimulationClock_
 
 instance Clock m GlossSimulationClock_ where
-  type TimeDomainOf GlossSimulationClock_ = ()
-  type Tag          GlossSimulationClock_ = Float
-  startClock _ = error errMsg
+  type Time GlossSimulationClock_ = ()
+  type Tag  GlossSimulationClock_ = Float
+  initClock _ = error errMsg
 
 -- | The clock that ticks for every @gloss@ simulation step.
 --   Use 'withProperSimClock' to transform to 'GlossSimulationClock_'.
 data GlossSimulationClock = GlossSimulationClock
 
 instance Clock m GlossSimulationClock where
-  type TimeDomainOf GlossSimulationClock = Float
-  type Tag          GlossSimulationClock = ()
-  startClock _ = error errMsg
+  type Time GlossSimulationClock = Float
+  type Tag  GlossSimulationClock = ()
+  initClock _ = error errMsg
 
--- | To use all features of the 'SyncSF' framework,
+-- | To use all features of the 'ClSF' framework,
 --   write your synchronous stream function on the 'GlossSimulationClock'
 --   and then use this function to transform it.
 withProperSimClock
   :: Monad m
-  => SyncSF m GlossSimulationClock  a b
-  -> SyncSF m GlossSimulationClock_ a b
-withProperSimClock syncsf = readerS
-  $ (intermingle *** Category.id) >>> runReaderS syncsf
+  => ClSF m GlossSimulationClock  a b
+  -> ClSF m GlossSimulationClock_ a b
+withProperSimClock clsf = readerS
+  $ (intermingle *** Category.id) >>> runReaderS clsf
   where
     intermingle :: Monad m => MSF m (TimeInfo GlossSimulationClock_) (TimeInfo GlossSimulationClock)
     intermingle = proc TimeInfo {tag} -> do
       let sinceTick = tag
       absolute <- sumS -< sinceTick
-      let sinceStart = absolute
+      let sinceInit = absolute
       returnA          -< TimeInfo { tag = (), .. }
 
 -- | The clock that ticks for every @gloss@ graphics output.
 data GlossGraphicsClock = GlossGraphicsClock
 
 instance Clock m GlossGraphicsClock where
-  type TimeDomainOf GlossGraphicsClock = ()
-  type Tag          GlossGraphicsClock = ()
-  startClock _ = error errMsg
+  type Time GlossGraphicsClock = ()
+  type Tag  GlossGraphicsClock = ()
+  initClock _ = error errMsg
 
 -- | A schedule you can't actually use, for internal purposes.
 glossSchedule :: Schedule Identity (SelectClock GlossEventClock a) GlossSimulationClock_