+++ rhine/src/FRP/Rhine/ClSF/Util.hs
@@ -4,52 +4,88 @@
 {-# LANGUAGE RecordWildCards  #-}
 {-# LANGUAGE TypeFamilies     #-}
 
-module FRP.Rhine.SyncSF.Util where
+module FRP.Rhine.ClSF.Util where
 
 
 -- base
 import Control.Arrow
 import Control.Category (Category)
 import qualified Control.Category (id)
+import Data.Maybe (fromJust)
+import Data.Monoid (Last (Last), getLast)
+
+-- containers
+import Data.Sequence
 
 -- transformers
 import Control.Monad.Trans.Reader (ask, asks)
 
 -- dunai
+import Control.Monad.Trans.MSF.Reader (readerS)
 import Data.MonadicStreamFunction (arrM_, sumFrom, delay, feedback)
+import Data.MonadicStreamFunction.Instances.VectorSpace ()
 import Data.VectorSpace
 
 -- rhine
-import FRP.Rhine.SyncSF.Core
-import FRP.Rhine.SyncSF.Except
+import FRP.Rhine.ClSF.Core
+import FRP.Rhine.ClSF.Except
 
 
 -- * Read time information
 
 -- | Read the environment variable, i.e. the 'TimeInfo'.
-timeInfo :: Monad m => SyncSF m cl a (TimeInfo cl)
+timeInfo :: Monad m => ClSF m cl a (TimeInfo cl)
 timeInfo = arrM_ ask
 
 {- | Utility to apply functions to the current 'TimeInfo',
 such as record selectors:
 @
-printAbsoluteTime :: SyncSF IO cl () ()
-printAbsoluteTime = timeInfoOf absolute >>> arrMSync print
+printAbsoluteTime :: ClSF IO cl () ()
+printAbsoluteTime = timeInfoOf absolute >>> arrMCl print
 @
 -}
-timeInfoOf :: Monad m => (TimeInfo cl -> b) -> SyncSF m cl a b
+timeInfoOf :: Monad m => (TimeInfo cl -> b) -> ClSF m cl a b
 timeInfoOf f = arrM_ $ asks f
 
--- | Calculate the time passed since the 'SyncSF' was instantiated.
-timeSinceSimStart :: (Monad m, TimeDomain td) => BehaviourF m td a (Diff td)
-timeSinceSimStart = proc _ -> do
-  time      <- timeInfoOf absolute -< ()
-  startTime <- keepFirst           -< time
-  returnA                          -< time `diffTime` startTime
+-- | Continuously return the time difference since the last tick.
+sinceTickS :: Monad m => ClSF m cl a (Diff (Time cl))
+sinceTickS = timeInfoOf sinceTick
+
+-- | Continuously return the time difference since clock initialisation.
+sinceInitS :: Monad m => ClSF m cl a (Diff (Time cl))
+sinceInitS = timeInfoOf sinceInit
+
+-- | Continuously return the absolute time.
+absoluteS :: Monad m => ClSF m cl a (Time cl)
+absoluteS = timeInfoOf absolute
 
 -- | Continuously return the tag of the current tick.
-theTag :: Monad m => SyncSF m cl a (Tag cl)
-theTag = timeInfoOf tag
+tagS :: Monad m => ClSF m cl a (Tag cl)
+tagS = timeInfoOf tag
+
+{- |
+Calculate the time passed since this 'ClSF' was instantiated.
+This is _not_ the same as 'sinceInitS',
+which measures the time since clock initialisation.
+
+For example, the following gives a sawtooth signal:
+
+@
+sawtooth = safely $ do
+  try $ sinceStart >>> proc time -> do
+    throwOn () -< time > 1
+    returnA    -< time
+  safe sawtooth
+@
+
+If you replace 'sinceStart' by 'sinceInitS',
+it will usually hang after one second,
+since it doesn't reset after restarting the sawtooth.
+-}
+sinceStart :: (Monad m, TimeDomain time) => BehaviourF m time a (Diff time)
+sinceStart = absoluteS >>> proc time -> do
+  startTime <- keepFirst -< time
+  returnA                -< time `diffTime` startTime
 
 
 -- * Useful aliases
@@ -58,11 +94,11 @@ theTag = timeInfoOf tag
 {- | Alias for 'Control.Category.>>>' (sequential composition)
 with higher operator precedence, designed to work with the other operators, e.g.:
 
-> syncsf1 >-> syncsf2 @@ clA **@ sched @** syncsf3 >-> syncsf4 @@ clB
+> clsf1 >-> clsf2 @@ clA ||@ sched @|| clsf3 >-> clsf4 @@ clB
 
 The type signature specialises e.g. to
 
-> (>->) :: Monad m => SyncSF m cl a b -> SyncSF m cl b c -> SyncSF m cl a c
+> (>->) :: Monad m => ClSF m cl a b -> ClSF m cl b c -> ClSF m cl a c
 -}
 infixr 6 >->
 (>->) :: Category cat
@@ -82,18 +118,20 @@ infixl 6 <-<
 {- | Output a constant value.
 Specialises e.g. to this type signature:
 
-> arr_ :: Monad m => b -> SyncSF m cl a b
+> arr_ :: Monad m => b -> ClSF m cl a b
 -}
 arr_ :: Arrow a => b -> a c b
 arr_ = arr . const
 
 
 -- | The identity synchronous stream function.
-syncId :: Monad m => SyncSF m cl a a
-syncId = Control.Category.id
+clId :: Monad m => ClSF m cl a a
+clId = Control.Category.id
 
 
 -- * Basic signal processing components
+
+-- ** Integration and differentiation
 
 -- | The output of @integralFrom v0@ is the numerical Euler integral
 --   of the input, with initial offset @v0@.
@@ -132,6 +170,28 @@ derivative
   => BehaviorF m td v v
 derivative = derivativeFrom zeroVector
 
+-- | Like 'derivativeFrom', but uses three samples to compute the derivative.
+--   Consequently, it is delayed by one sample.
+threePointDerivativeFrom
+  :: ( Monad m, VectorSpace v
+     , Groundfield v ~ Diff td)
+  => v -- ^ The initial position
+  -> BehaviorF m td v v
+threePointDerivativeFrom v0 = proc v -> do
+  dv  <- derivativeFrom v0 -< v
+  dv' <- delay zeroVector  -< dv
+  returnA                  -< (dv ^+^ dv') ^/ 2
+
+-- | Like 'threePointDerivativeFrom',
+--   but with the initial position initialised to 'zeroVector'.
+threePointDerivative
+  :: ( Monad m, VectorSpace v
+     , Groundfield v ~ Diff td)
+  => BehaviorF m td v v
+threePointDerivative = threePointDerivativeFrom zeroVector
+
+-- ** Averaging and filters
+
 -- | A weighted moving average signal function.
 --   The output is the average of the first input,
 --   weighted by the second input
@@ -151,7 +211,8 @@ weightedAverageFrom v0 = feedback v0 $ proc ((v, weight), vAvg) -> do
 
 -- | An exponential moving average, or low pass.
 --   It will average out, or filter,
---   all features below a given time scale.
+--   all features below a given time constant @t@.
+--   (Equivalently, it filters out frequencies above @1 / (2 * pi * t)@.)
 averageFrom
   :: ( Monad m, VectorSpace v
      , Floating (Groundfield v)
@@ -199,11 +260,75 @@ averageLin
   -> BehaviourF m td v v
 averageLin = averageLinFrom zeroVector
 
+-- *** First-order filters
+
+-- | Alias for 'average'.
+lowPass
+  :: ( Monad m, VectorSpace v
+     , Floating (Groundfield v)
+     , Groundfield v ~ Diff td)
+  => Diff td
+  -> BehaviourF m td v v
+lowPass = average
+
+-- | Filters out frequencies below @1 / (2 * pi * t)@.
+highPass
+  :: ( Monad m, VectorSpace v
+     , Floating (Groundfield v)
+     , Groundfield v ~ Diff td)
+  => Diff td -- ^ The time constant @t@
+  -> BehaviourF m td v v
+highPass t = clId ^-^ lowPass t
+
+-- | Filters out frequencies other than @1 / (2 * pi * t)@.
+bandPass
+  :: ( Monad m, VectorSpace v
+     , Floating (Groundfield v)
+     , Groundfield v ~ Diff td)
+  => Diff td -- ^ The time constant @t@
+  -> BehaviourF m td v v
+bandPass t = lowPass t >>> highPass t
+
+-- | Filters out the frequency @1 / (2 * pi * t)@.
+bandStop
+  :: ( Monad m, VectorSpace v
+     , Floating (Groundfield v)
+     , Groundfield v ~ Diff td)
+  => Diff td -- ^ The time constant @t@
+  -> BehaviourF m td v v
+bandStop t = clId ^-^ bandPass t
+
+
+
+-- * Delays
+
 -- | Remembers and indefinitely outputs ("holds") the first input value.
-keepFirst :: Monad m => SyncSF m cl a a
+keepFirst :: Monad m => ClSF m cl a a
 keepFirst = safely $ do
   a <- try throwS
   safe $ arr $ const a
+
+-- | Remembers all input values that arrived within a given time window.
+--   New values are appended left.
+historySince
+  :: (Monad m, Ord (Diff (Time cl)), TimeDomain (Time cl))
+  => Diff (Time cl) -- ^ The size of the time window
+  -> ClSF m cl a (Seq (TimeInfo cl, a))
+historySince dTime = readerS $ accumulateWith appendValue empty
+  where
+    appendValue (ti, a) tias  = takeWhileL (recentlySince ti) $ (ti, a) <| tias
+    recentlySince ti (ti', _) = diffTime (absolute ti) (absolute ti') < dTime
+
+-- | Delay a signal by certain time span.
+delayBy
+  :: (Monad m, Ord (Diff (Time cl)), TimeDomain (Time cl))
+  => Diff (Time cl) -- ^ The time span to delay the signal
+  -> ClSF m cl a a
+delayBy dTime = historySince dTime >>> arr (viewr >>> safeHead) >>> lastS undefined >>> arr snd
+  where
+    safeHead EmptyR   = Nothing
+    safeHead (_ :> a) = Just a
+-- TODO Think about how to do it without undefined (maybe exceptions)
 
 -- * Timers
 
@@ -217,9 +342,9 @@ timer
   => Diff td
   -> BehaviorF (ExceptT () m) td a (Diff td)
 timer diff = proc _ -> do
-  sinceSimStart <- timeSinceSimStart -< ()
-  _             <- throwOn ()        -< sinceSimStart > diff
-  returnA                            -< sinceSimStart
+  time <- sinceStart -< ()
+  _    <- throwOn () -< time > diff
+  returnA            -< time
 
 -- | Like 'timer_', but doesn't output the remaining time at all.
 timer_
@@ -241,3 +366,11 @@ scaledTimer
   => Diff td
   -> BehaviorF (ExceptT () m) td a (Diff td)
 scaledTimer diff = timer diff >>> arr (/ diff)
+
+
+-- * To be ported to Dunai
+
+-- | Remembers the last 'Just' value,
+--   defaulting to the given initialisation value.
+lastS :: Monad m => a -> MSF m (Maybe a) a
+lastS a = arr Last >>> mappendFrom (Last (Just a)) >>> arr (getLast >>> fromJust)