+++ rhine/src/FRP/Rhine/Clock/Realtime/Audio.hs
@@ -12,6 +12,7 @@ module FRP.Rhine.Clock.Realtime.Audio
   ( AudioClock (..)
   , AudioRate (..)
   , PureAudioClock (..)
+  , PureAudioClockF
   , pureAudioClockF
   )
   where
@@ -89,10 +90,10 @@ theBufferSize = fromInteger . natVal
 
 instance (MonadIO m, KnownNat bufferSize, AudioClockRate rate)
       => Clock m (AudioClock rate bufferSize) where
-  type TimeDomainOf (AudioClock rate bufferSize) = UTCTime
-  type Tag          (AudioClock rate bufferSize) = Maybe Double
+  type Time (AudioClock rate bufferSize) = UTCTime
+  type Tag  (AudioClock rate bufferSize) = Maybe Double
 
-  startClock audioClock = do
+  initClock audioClock = do
     let
       step       = picosecondsToDiffTime -- The only sufficiently precise conversion function
                      $ round (10 ^ (12 :: Integer) / theRateNum audioClock :: Double)
@@ -135,10 +136,10 @@ class PureAudioClockRate (rate :: AudioRate) where
 
 
 instance (Monad m, PureAudioClockRate rate) => Clock m (PureAudioClock rate) where
-  type TimeDomainOf (PureAudioClock rate) = Double
-  type Tag          (PureAudioClock rate) = ()
+  type Time (PureAudioClock rate) = Double
+  type Tag  (PureAudioClock rate) = ()
 
-  startClock audioClock = return
+  initClock audioClock = return
     ( arr (const (1 / thePureRateNum audioClock)) >>> sumS &&& arr (const ())
     , 0
     )