+++ rhine/src/FRP/Rhine/Clock/Realtime/Busy.hs
@@ -16,10 +16,10 @@ side effects, time measurement and framework overhead.
 data Busy = Busy
 
 instance Clock IO Busy where
-  type TimeDomainOf Busy = UTCTime
-  type Tag          Busy = ()
+  type Time Busy = UTCTime
+  type Tag  Busy = ()
 
-  startClock _ = do
+  initClock _ = do
     initialTime <- getCurrentTime
     return
       ( arrM_ getCurrentTime