+++ rhine/src/FRP/Rhine/ResamplingBuffer/KeepLast.hs
@@ -6,6 +6,8 @@ import FRP.Rhine.ResamplingBuffer.Timeless
 
 -- | Always keeps the last input value,
 --   or in case of no input an initialisation value.
+--   If @cl2@ approximates continuity,
+--   this behaves like a zero-order hold.
 keepLast :: Monad m => a -> ResamplingBuffer m cl1 cl2 a a
 keepLast = timelessResamplingBuffer AsyncMealy {..}
   where